requirements.txt
psycopg2-binary==2.8.6
streamlit==0.84.2
sqlalchemy==1.3.23
watchdog==2.1.3